fix: menus.yaml && dashboard
